Update generated files with build 70488

+++ me/teamwork/sendactivitynotification/send_activity_notification_request_bodyable.go
@@ -10,12 +10,12 @@ type SendActivityNotificationRequestBodyable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetActivityType()(*string)
-    GetChainId()(*int64)
+    GetChainId()(*int32)
     GetPreviewText()(i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.ItemBodyable)
     GetTemplateParameters()([]i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.KeyValuePairable)
     GetTopic()(i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.TeamworkActivityTopicable)
     SetActivityType(value *string)()
-    SetChainId(value *int64)()
+    SetChainId(value *int32)()
     SetPreviewText(value i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.ItemBodyable)()
     SetTemplateParameters(value []i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.KeyValuePairable)()
     SetTopic(value i4a838ef194e4c99e9f2c63ba10dab9cb120a89367c1d4ab0daa63bb424e20d87.TeamworkActivityTopicable)()
