+++ models/microsoft/graph/drive_item_version.go
@@ -10,7 +10,7 @@ type DriveItemVersion struct {
     // The content stream for this version of the item.
     content []byte;
     // Indicates the size of the content stream for this version of the item.
-    size *int64;
+    size *int32;
 }
 // NewDriveItemVersion instantiates a new driveItemVersion and sets the default values.
 func NewDriveItemVersion()(*DriveItemVersion) {
@@ -45,7 +45,7 @@ func (m *DriveItemVersion) GetFieldDeserializers()(map[string]func(interface{}, 
         return nil
     }
     res["size"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -57,7 +57,7 @@ func (m *DriveItemVersion) GetFieldDeserializers()(map[string]func(interface{}, 
     return res
 }
 // GetSize gets the size property value. Indicates the size of the content stream for this version of the item.
-func (m *DriveItemVersion) GetSize()(*int64) {
+func (m *DriveItemVersion) GetSize()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -80,7 +80,7 @@ func (m *DriveItemVersion) Serialize(writer i04eb5309aeaafadd28374d79c8471df9b26
         }
     }
     {
-        err = writer.WriteInt64Value("size", m.GetSize())
+        err = writer.WriteInt32Value("size", m.GetSize())
         if err != nil {
             return err
         }
@@ -94,7 +94,7 @@ func (m *DriveItemVersion) SetContent(value []byte)() {
     }
 }
 // SetSize sets the size property value. Indicates the size of the content stream for this version of the item.
-func (m *DriveItemVersion) SetSize(value *int64)() {
+func (m *DriveItemVersion) SetSize(value *int32)() {
     if m != nil {
         m.size = value
     }
