+++ models/microsoft/graph/print_usageable.go
@@ -8,12 +8,12 @@ import (
 type PrintUsageable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetCompletedBlackAndWhiteJobCount()(*int64)
-    GetCompletedColorJobCount()(*int64)
-    GetIncompleteJobCount()(*int64)
+    GetCompletedBlackAndWhiteJobCount()(*int32)
+    GetCompletedColorJobCount()(*int32)
+    GetIncompleteJobCount()(*int32)
     GetUsageDate()(*i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.DateOnly)
-    SetCompletedBlackAndWhiteJobCount(value *int64)()
-    SetCompletedColorJobCount(value *int64)()
-    SetIncompleteJobCount(value *int64)()
+    SetCompletedBlackAndWhiteJobCount(value *int32)()
+    SetCompletedColorJobCount(value *int32)()
+    SetIncompleteJobCount(value *int32)()
     SetUsageDate(value *i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.DateOnly)()
 }
