+++ models/microsoft/graph/quotaable.go
@@ -8,16 +8,16 @@ import (
 type Quotaable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetDeleted()(*int64)
-    GetRemaining()(*int64)
+    GetDeleted()(*int32)
+    GetRemaining()(*int32)
     GetState()(*string)
     GetStoragePlanInformation()(StoragePlanInformationable)
-    GetTotal()(*int64)
-    GetUsed()(*int64)
-    SetDeleted(value *int64)()
-    SetRemaining(value *int64)()
+    GetTotal()(*int32)
+    GetUsed()(*int32)
+    SetDeleted(value *int32)()
+    SetRemaining(value *int32)()
     SetState(value *string)()
     SetStoragePlanInformation(value StoragePlanInformationable)()
-    SetTotal(value *int64)()
-    SetUsed(value *int64)()
+    SetTotal(value *int32)()
+    SetUsed(value *int32)()
 }
